separated files in View
