+++ RamenShopApp/View/Admin/ReloadRequestsButton.swift
@@ -1,0 +1,24 @@
+//
+//  ReloadRequestsButton.swift
+//  RamenShopApp
+//
+//  Created by Koro Saka on 2021-02-24.
+//  Copyright © 2021 Koro Saka. All rights reserved.
+//
+
+import SwiftUI
+
+struct ReloadRequestsButton: View {
+    @ObservedObject var viewModel: AdminPageViewModel
+    var body: some View {
+        Button(action: {
+            viewModel.fetchRequests()
+        }) {
+            Text("reload")
+                .containingSymbol(symbol: "arrow.triangle.2.circlepath",
+                                  color: .strongPink,
+                                  textFont: .title3,
+                                  symbolFont: .title3)
+        }
+    }
+}
