+++ RamenShopApp/View/Common/PickerRole.swift
@@ -1,0 +1,13 @@
+//
+//  PickerRole.swift
+//  RamenShopApp
+//
+//  Created by Koro Saka on 2021-02-24.
+//  Copyright © 2021 Koro Saka. All rights reserved.
+//
+
+import Foundation
+
+enum PickerRole {
+    case reviewPicture, userIcon
+}
