+++ RamenShopApp/View/GoogleMap/MapType.swift
@@ -1,0 +1,13 @@
+//
+//  MapType.swift
+//  RamenShopApp
+//
+//  Created by Koro Saka on 2021-02-24.
+//  Copyright © 2021 Koro Saka. All rights reserved.
+//
+
+import Foundation
+
+enum MapType {
+    case searching, registering, admin, fromShop
+}
