+++ RamenShopApp/View/Profile/FavoriteShopInfo.swift
@@ -1,0 +1,16 @@
+//
+//  FavoriteShopInfo.swift
+//  RamenShopApp
+//
+//  Created by Koro Saka on 2021-02-24.
+//  Copyright © 2021 Koro Saka. All rights reserved.
+//
+
+import SwiftUI
+
+
+struct FavoriteShopInfo: Identifiable {
+    let id: String //MARK: ShopID
+    var shopName: String?
+    var shopTopImage: Image?
+}
