+++ RamenShopApp/View/Request/RequestStatusView.swift
@@ -52,121 +52,3 @@ struct RequestStatusView: View {
         .navigationBarHidden(true)
     }
 }
-
-struct RequestInfo: View {
-    @EnvironmentObject var viewModel: RequestStatusViewModel
-    var body: some View {
-        VStack(spacing: 0) {
-            Spacer().frame(height: 10).wideStyle()
-            RequestedShopName(name: viewModel.shopName ?? Constants.EMPTY)
-            Spacer().frame(height: 30)
-            ReviewStatus()
-            Spacer().frame(height: 10)
-        }
-        .background(Color.superWhitePasteGreen)
-        .cornerRadius(20)
-    }
-}
-
-struct RequestedShopName: View {
-    let name: String
-    var body: some View {
-        VStack {
-            Text("shop name")
-                .foregroundColor(.black)
-                .underline()
-            Spacer().frame(height: 5)
-            Text(name)
-                .largestTitleStyleWithColor(color: .viridianGreen)
-        }
-    }
-}
-
-struct ReviewStatus: View {
-    @EnvironmentObject var viewModel: RequestStatusViewModel
-    var body: some View {
-        VStack {
-            Text("review status")
-                .foregroundColor(.black)
-                .underline()
-            Spacer().frame(height: 5)
-            Text(viewModel.inspectionStatus?.getStatus() ?? Constants.EMPTY).largestTitleStyleWithColor(color: viewModel.inspectionStatus?.getStatusColor() ?? .black)
-            Spacer().frame(height: 5)
-            Text(viewModel.inspectionStatus?.getSubMessage() ?? Constants.EMPTY).foregroundColor(.gray)
-            Spacer().frame(height: 30)
-            if viewModel.isRejected {
-                RejectReason()
-                    .padding(5)
-                Spacer().frame(height: 30)
-            }
-        }
-    }
-}
-
-struct RemoveButton: View {
-    @EnvironmentObject var viewModel: RequestStatusViewModel
-    var body: some View {
-        Button(action: {
-            viewModel.isShowAlert = true
-        }) {
-            Text(viewModel.inspectionStatus?.getButtonMessage() ?? Constants.EMPTY)
-                .containingSymbolWide(symbol: "trash",
-                                      color: .strongRed,
-                                      textFont: .title,
-                                      symbolFont: .title2)
-        }
-        .alert(isPresented: $viewModel.isShowAlert) {
-            switch viewModel.activeAlert {
-            case .confirmation:
-                return Alert(title: Text("Confirmation"),
-                             message: Text(viewModel.inspectionStatus?.getConfirmationMessage() ?? Constants.EMPTY),
-                             primaryButton: .default(Text("Yes")) {
-                                viewModel.onClickConfirmation()
-                             },
-                             secondaryButton: .cancel(Text("cancel")))
-            case .completion:
-                return Alert(title: Text("Success"),
-                             message: Text("Updating data was succeeded"),
-                             dismissButton: .default(Text("Close")) {
-                                viewModel.resetData()
-                             })
-            case .error:
-                return Alert(title: Text("Fail"),
-                             message: Text("Updating data was failed"),
-                             dismissButton: .default(Text("Close")) {
-                                viewModel.resetData()
-                             })
-            }
-        }
-    }
-}
-
-struct RejectReason: View {
-    @EnvironmentObject var viewModel: RequestStatusViewModel
-    var body: some View {
-        VStack {
-            Text("reason for reject")
-                .foregroundColor(.black)
-                .underline()
-                .wideStyle()
-            Spacer().frame(height: 5)
-            Text(viewModel.rejectReason)
-                .foregroundColor(.navy)
-                .bold()
-        }
-    }
-}
-
-struct ReloadStatusButton: View {
-    @EnvironmentObject var viewModel: RequestStatusViewModel
-    var body: some View {
-        Button(action: {
-            self.viewModel.fetchRequestStatus()
-        }) {
-            Image(systemName: "arrow.triangle.2.circlepath")
-                .circleSymbol(font: .title3,
-                              fore: .white,
-                              back: .strongPink)
-        }
-    }
-}
