+++ RamenShopApp/View/ShopSearch/Shop/LatestReviews.swift
@@ -1,0 +1,61 @@
+//
+//  LatestReviews.swift
+//  RamenShopApp
+//
+//  Created by Koro Saka on 2021-02-24.
+//  Copyright © 2021 Koro Saka. All rights reserved.
+//
+
+import SwiftUI
+
+struct LatestReviews: View {
+    let latestReviews: [Review]
+    let shop: Shop?
+    
+    var body: some View {
+        Text("latest reviews")
+            .foregroundColor(.white)
+            .wideStyle()
+            .upDownPadding(size: 3)
+            .background(Color.gray)
+            .shadow(color: .black, radius: 1)
+        if latestReviews.count == 0 {
+            Spacer().frame(height: 10)
+            Text("there is no review")
+                .foregroundColor(.viridianGreen)
+                .shadow(color: .black, radius: 0.5, x: 0.5, y: 0.5)
+        }
+        
+        if latestReviews.count > 0 {
+            Spacer().frame(height: 10)
+            ReviewHeadline(viewModel: .init(review: latestReviews[0]))
+                .padding(.init(top: 0,
+                               leading: 0,
+                               bottom: 0,
+                               trailing: 15))
+        }
+        if latestReviews.count > 1 {
+            Spacer().frame(height: 10)
+            ReviewHeadline(viewModel: .init(review: latestReviews[1]))
+                .padding(.init(top: 10,
+                               leading: 0,
+                               bottom: 0,
+                               trailing: 15))
+        }
+        if latestReviews.count > 0 {
+            Spacer().frame(height: 20)
+            HStack {
+                Spacer()
+                if let _shop = shop {
+                    NavigationLink(destination: AllReviewView(viewModel: .init(shop: _shop))) {
+                        Text("all review...")
+                            .foregroundColor(.seaBlue)
+                            .underline()
+                            .sidePadding(size: 15)
+                    }
+                }
+            }
+        }
+    }
+}
+
