+++ RamenShopApp/View/ShopSearch/ShopsMapView.swift
@@ -45,38 +45,3 @@ struct ShopsMapView: View {
         }
     }
 }
-
-struct EvaluationFilter: View {
-    @ObservedObject var viewModel: ShopsMapViewModel
-    
-    var body: some View {
-        VStack(spacing: 0) {
-            Text("evaluation filter")
-                .foregroundColor(.viridianGreen)
-                .shadow(color: .black, radius: 0.5, x: 0.5, y: 0.5)
-            Picker(selection: $viewModel.evaluationFilter, label: Text("EvaFilter"), content: {
-                Text("---").tag(-1)
-                ForEach(0..<viewModel.filterValues.count, id: \.self) { index in
-                    Text(String(viewModel.filterValues[index])).tag(index)
-                }
-            })
-            .pickerStyle(SegmentedPickerStyle())
-        }
-    }
-}
-
-//MARK: TODO
-//struct ReloadButton: View {
-//    let buttonAction: ()
-//    var body: some View {
-//        Button(action: {
-//            buttonAction
-//        }) {
-//            Text("reload")
-//                .containingSymbol(symbol: "arrow.triangle.2.circlepath",
-//                                  color: .strongPink,
-//                                  textFont: .title3,
-//                                  symbolFont: .title3)
-//        }
-//    }
-//}
